+++ exps/NWA12.COBALT/driver.sh
@@ -9,6 +9,19 @@
 #SBATCH --clusters=c6
 #SBATCH --account=ira-cefi
 
+# Default to not using shared project folders
+USE_PROJ_SHARED=false
+
+# Parse optional argument
+for arg in "$@"; do
+  case $arg in
+    --use-proj-shared)
+      USE_PROJ_SHARED=true
+      shift
+      ;;
+  esac
+done
+
 #
 ntasks1=1600
 ntasks2=900
@@ -39,10 +52,12 @@ echo "SET APPTAINER_BIND"
 export APPTAINER_BIND="/usr/share/libdrm,/var/spool/slurmd,/opt/cray,/opt/intel,${PWD},/etc/libibverbs.d,/usr/lib64/libibverbs,/usr/lib64/libnl3-200,${HOME}"
 
 #
-echo "clean RESTART folders ..."
-rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_48hrs/*
-rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs/*
-rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs_rst/*
+if $USE_PROJ_SHARED; then
+  echo "clean RESTART folders ..."
+  rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_48hrs/*
+  rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs/*
+  rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs_rst/*
+fi
 
 echo "run 40x40 48hrs test ..."
 ln -fs input.nml_48hr input.nml
@@ -50,7 +65,9 @@ pushd INPUT/
 ln -fs MOM_layout_40 MOM_layout
 ln -fs MOM_layout_40 SIS_layout
 popd
-ln -fs /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_48hrs ./RESTART
+if $USE_PROJ_SHARED; then
+  ln -fs /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_48hrs ./RESTART
+fi
 srun --ntasks ${ntasks1} --export=ALL apptainer exec --writable-tmpfs $img ./execrunscript.sh > out1 2>err1
 mv RESTART RESTART_48hrs
 mv ocean.stats RESTART_48hrs
@@ -58,7 +75,9 @@ mv ocean.stats RESTART_48hrs
 #
 echo "run 40x40 24hrs test ..."
 ln -fs input.nml_24hr input.nml
-ln -fs /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs ./RESTART
+if $USE_PROJ_SHARED; then
+  ln -fs /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs ./RESTART
+fi
 srun --ntasks ${ntasks1} --export=ALL apptainer exec --writable-tmpfs $img ./execrunscript.sh > out2 2>err2
 mv RESTART RESTART_24hrs
 mv ocean.stats RESTART_24hrs
@@ -76,7 +95,9 @@ pushd INPUT/
 ln -fs MOM_layout_30 MOM_layout
 ln -fs MOM_layout_30 SIS_layout
 popd
-ln -fs /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs_rst ./RESTART
+if $USE_PROJ_SHARED; then
+  ln -fs /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs_rst ./RESTART
+fi
 srun --ntasks ${ntasks2} --export=ALL apptainer exec --writable-tmpfs $img ./execrunscript.sh > out3 2>err3
 mv RESTART RESTART_24hrs_rst
 mv ocean.stats RESTART_24hrs_rst
@@ -102,9 +123,11 @@ done
 echo "All restart files are identical, PASS"
 
 #
-echo "clean RESTART folders now ..."
-rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_48hrs/*
-rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs/*
-rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs_rst/*
+if $USE_PROJ_SHARED; then
+  echo "clean RESTART folders now ..."
+  rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_48hrs/*
+  rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs/*
+  rm -rf /gpfs/f6/ira-cefi/proj-shared/github/tmp/NWA12/RESTART_24hrs_rst/*
+fi
 
 echo "Test ended:  " `date`
